# docs/openapi-mvp-skeleton-draft.yaml
openapi: 3.1.0
info:
  title: Order System v2 API
  version: 0.1.0
paths:
...
  /api/v1/order-item-allocations:
    get:
      tags:
      - order-item-allocations
      summary: List Order Item Allocation Work Items
      operationId: list_order_item_allocation_work_items_api_v1_order_item_allocations_get
      security:
      - HTTPBearer: []
      parameters:
      - name: unallocated_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Unallocated Only
      - name: delivery_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Delivery Date
      - name: order_status
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              $ref: '#/components/schemas/OrderStatus'
          - type: 'null'
          title: Order Status
      - name: supplier_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            exclusiveMinimum: 0
          - type: 'null'
          title: Supplier Id
      - name: product_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 255
          - type: 'null'
          title: Product Name
      - name: customer_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 255
          - type: 'null'
          title: Customer Name
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 500
            minimum: 1
          - type: 'null'
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderItemAllocationWorkItem'
                title: Response List Order Item Allocation Work Items Api V1 Order
                  Item Allocations Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/order-item-allocations/suggestions:
    post:
      tags:
      - order-item-allocations
      summary: Suggest Allocations
      operationId: suggest_allocations_api_v1_order_item_allocations_suggestions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllocationSuggestRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AllocationSuggestion'
                type: array
                title: Response Suggest Allocations Api V1 Order Item Allocations
                  Suggestions Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
      - HTTPBearer: []
  /api/v1/order-item-allocations/bulk-save:
    post:
      tags:
      - order-item-allocations
      summary: Bulk Save Allocations
      operationId: bulk_save_allocations_api_v1_order_item_allocations_bulk_save_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkAllocationSaveRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkAllocationSaveResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
      - HTTPBearer: []
  /api/v1/orders:
    get:
      tags:
      - orders
      summary: List Orders
      operationId: list_orders_api_v1_orders_get
      security:
      - HTTPBearer: []
      parameters:
      - name: stale_delivery_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Stale Delivery Only
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
                title: Response List Orders Api V1 Orders Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - orders
      summary: Create Order
      operationId: create_order_api_v1_orders_post
      deprecated: true
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/orders/{order_id}:
    get:
      tags:
      - orders
      summary: Get Order
      operationId: get_order_api_v1_orders__order_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    patch:
      tags:
      - orders
      summary: Update Order
      operationId: update_order_api_v1_orders__order_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          title: Order Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /api/v1/orders/{order_id}/items:
    get:
      tags:
      - orders
      summary: List Order Items
      operationId: list_order_items_api_v1_orders__order_id__items_get
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderItemResponse'
                title: Response List Order Items Api V1 Orders  Order Id  Items Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - orders
      summary: Create Order Item
      operationId: create_order_item_api_v1_orders__order_id__items_post
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          title: Order Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItemCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderItemResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/orders/uuid/{order_uuid}/items:
    get:
      tags:
      - orders
      summary: List Order Items By Order Uuid
      operationId: list_order_items_by_order_uuid_api_v1_orders_uuid__order_uuid__items_get
      security:
      - HTTPBearer: []
      parameters:
      - name: order_uuid
        in: path
        required: true
        schema:
          type: string
          title: Order Uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderItemResponse'
                title: Response List Order Items By Order Uuid Api V1 Orders Uuid  Order
                  Uuid  Items Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - orders
      summary: Create Order Item By Order Uuid
      operationId: create_order_item_by_order_uuid_api_v1_orders_uuid__order_uuid__items_post
      security:
      - HTTPBearer: []
      parameters:
      - name: order_uuid
        in: path
        required: true
        schema:
          type: string
          title: Order Uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItemCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderItemResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/orders/{order_id}/items/{item_id}:
    patch:
      tags:
      - orders
      summary: Update Order Item
      operationId: update_order_item_api_v1_orders__order_id__items__item_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          title: Order Id
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          title: Item Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItemUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderItemResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - orders
      summary: Delete Order Item
      operationId: delete_order_item_api_v1_orders__order_id__items__item_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          title: Order Id
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          title: Item Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/orders/uuid/{order_uuid}/items/{item_uuid}:
    patch:
      tags:
      - orders
      summary: Update Order Item By Uuid
      operationId: update_order_item_by_uuid_api_v1_orders_uuid__order_uuid__items__item_uuid__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: order_uuid
        in: path
        required: true
        schema:
          type: string
          title: Order Uuid
      - name: item_uuid
        in: path
        required: true
        schema:
          type: string
          title: Item Uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItemUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderItemResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - orders
      summary: Delete Order Item By Uuid
      operationId: delete_order_item_by_uuid_api_v1_orders_uuid__order_uuid__items__item_uuid__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: order_uuid
        in: path
        required: true
        schema:
          type: string
          title: Order Uuid
      - name: item_uuid
        in: path
        required: true
        schema:
          type: string
          title: Item Uuid
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/invoices/{invoice_id}/items:
    get:
      tags:
      - invoices
      summary: List Invoice Items
      operationId: list_invoice_items_api_v1_invoices__invoice_id__items_get
      security:
      - HTTPBearer: []
      parameters:
      - name: invoice_id
        in: path
        required: true
        schema:
          type: string
          title: Invoice Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceItemResponse'
                title: Response List Invoice Items Api V1 Invoices  Invoice Id  Items
                  Get
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/invoices/uuid/{invoice_uuid}/items:
    get:
      tags:
      - invoices
      summary: List Invoice Items By Uuid
      operationId: list_invoice_items_by_uuid_api_v1_invoices_uuid__invoice_uuid__items_get
      security:
      - HTTPBearer: []
      parameters:
      - name: invoice_uuid
        in: path
        required: true
        schema:
          type: string
          title: Invoice Uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceItemResponse'
                title: Response List Invoice Items By Uuid Api V1 Invoices Uuid  Invoice
                  Uuid  Items Get
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/invoices/{invoice_id}/items/{invoice_item_id}:
    patch:
      tags:
      - invoices
      summary: Update Invoice Draft Item
      operationId: update_invoice_draft_item_api_v1_invoices__invoice_id__items__invoice_item_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: invoice_id
        in: path
        required: true
        schema:
          type: string
          title: Invoice Id
      - name: invoice_item_id
        in: path
        required: true
        schema:
          type: string
          title: Invoice Item Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceItemUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceItemResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/invoices/uuid/{invoice_uuid}/items/{invoice_item_uuid}:
    patch:
      tags:
      - invoices
      summary: Update Invoice Draft Item By Uuid
      operationId: update_invoice_draft_item_by_uuid_api_v1_invoices_uuid__invoice_uuid__items__invoice_item_uuid__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: invoice_uuid
        in: path
        required: true
        schema:
          type: string
          title: Invoice Uuid
      - name: invoice_item_uuid
        in: path
        required: true
        schema:
          type: string
          title: Invoice Item Uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceItemUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceItemResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
components:
  schemas:
...
    AllocationSuggestRequest:
      properties:
        order_item_ids:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          maxItems: 500
          minItems: 1
          title: Order Item Ids
      type: object
      required:
      - order_item_ids
      title: AllocationSuggestRequest
...
    ApiErrorResponse:
      properties:
        detail:
          $ref: '#/components/schemas/ApiErrorDetail'
      type: object
      required:
      - detail
      title: ApiErrorResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OrderItemAllocationWorkItem:
      properties:
        order_item_id:
          type: string
          title: Order Item Id
        allocation_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Allocation Id
        order_id:
          type: string
          title: Order Id
        order_no:
          type: string
          title: Order No
        order_status:
          $ref: '#/components/schemas/OrderStatus'
        customer_name:
          type: string
          title: Customer Name
        product_id:
          type: integer
          title: Product Id
        product_name:
          type: string
          title: Product Name
        ordered_qty:
          type: number
          title: Ordered Qty
        delivery_date:
          type: string
          format: date
          title: Delivery Date
        shipped_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Shipped Date
        allocation_status:
          type: string
          title: Allocation Status
        allocated_supplier_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Allocated Supplier Id
        allocated_qty:
          anyOf:
          - type: number
          - type: 'null'
          title: Allocated Qty
      type: object
      required:
      - order_item_id
      - order_id
      - order_no
      - order_status
      - customer_name
      - product_id
      - product_name
      - ordered_qty
      - delivery_date
      - allocation_status
      title: OrderItemAllocationWorkItem
...
    OrderStatus:
      type: string
      enum:
      - new
      - confirmed
      - allocated
      - purchased
      - shipped
      - invoiced
      - cancelled
      title: OrderStatus
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer